Confirm that all KAS have observed our OAuth metadata
This change updates the well known check to verify that all masters
have observed our metadata config.  The IP of the masters is
determined by their endpoints resource.

This prevents a race in CI where the install will report completed
and start running tests that fail with an "Unauthorized" error.
This error occurs because a request was sent to a master that had
not observed our OAuth metadata and thus had the logic the honors
OAuth tokens disabled.

Bug 1694878

Signed-off-by: Monis Khan <mkhan@redhat.com>